added class to parse entries as hash
