+++ lib/google/taxonomy.rb
@@ -1,5 +1,6 @@
 require "google/taxonomy/version"
 require "google/taxonomy/validator"
+require "google/taxonomy/parser"
 
 module Google
   module Taxonomy
