+++ lib/google/taxonomy/parser.rb
@@ -1,0 +1,30 @@
+module Google
+  module Taxonomy
+    class Parser < Validator
+
+      def self.hash(lang = 'de-DE')
+        @@hash ||= hashed_entries(lang)
+      end
+
+      def self.hashed_entries(lang = 'de-DE')
+        split_entries = entries(lang).map {|e| e.split(/ > /)}
+        split_entries.reduce({}) do |hash, entry|
+          hash.deep_merge(entry.reverse.inject({}) {|k, v| {v => k}})
+        end
+      end
+
+      def self.get_next_level(categories = [], hash = hash())
+        if categories.blank?
+          return hash
+        else
+          next_level = hash["#{categories[0]}"]
+          if next_level.blank?
+            return hash
+          else
+            get_next_level(categories.drop(1), next_level)
+          end
+        end
+      end
+    end
+  end
+end
